+++ src/test/kotlin/com/jetbrains/python/psi/PyClass.kt
@@ -1,0 +1,6 @@
+package com.jetbrains.python.psi
+
+interface PyClass : PyElement {
+    fun getName(): String?
+    fun getQualifiedName(): String?
+}